Feat(UserManager, ChatServer): split user manager from client manager
